[catch2] Fix share/catch2 directory name conflict (#17271)
Catch2 installs into share/Catch2/ and lib/cmake/Catch2.
The vcpkg_fixup_cmake_targets() function tries to move the contents from
the latter directory to share/catch2, i.e. differing in case w.r.t. the
former.

On Linux this results in both names being present. On pure Windows the
first name "wins". It breaks in a Linux Docker container where the
installation root is located on a bound directory that is on a
case-insensitive file system, such as from a host running Windows.

Fixes: #17228.