[vcpkg_execute_required_process] check log file existence before reading
Check for the output log file existence before trying to read it. This
edge case was hit in this scenario: if the command critically fails
(e.g. the executable doesn't exist), the stdout file may be missing.
When this occurs, CMake stops prematurely due to not being able to
read the missing file. Desired behavior instead is to have this output
its error message on process failure.