[vcpkg_from_gitlab] Download archive from API
The GitLab API endpoint for downloading a repository archive supports
the Authorization header, whereas the website-based repository archive
URL only uses Cookie headers for authorization.

Because the API endpoint is also accessible with no authorization
headers when the repository is public, this URL should work for existing
projects that use `vcpkg_from_gitlab`.