[stb] Add support for stb_vorbis and update port.
I wanted to use the stb_vorbis library from stb.
The current stb port does not have support for the stb_vorbis library.

I have added support for stb_vorbis.

I have also added some updates to the port.
I have replaced the CONTROL file with a `vcpkg.json` manifest file.
In the replacement file, I have updated the version string to
"2021-06-20". That is the data of the latest commit in the stb
repository.
I updated the portfile's `REF` to the latest commit as well.

---

To add stb_vorbis I used the compilation guide at
https://nothings.org/stb_vorbis/.

I made the portfile first cut the header out from the source file and
write it to a seperate file.
Then, the portfile needs to compile the source file as a library.
First, it adds an export macro to the source file along with an include
to find the header.
Then it copies a `CMakeLists.txt` to the source path. I wrote this file
to configure and install a library for stb_vorbis.
Next, vcpkg configures and installs the library.

The portfile installs the rest of the stb libraries as normal.

I have no clue what `vcpkg-cmake-wrapper.cmake` is, and I do not think
`FindStb.cmake` is still needed. If I do not install them, I can still
use the other libraries, but I have kept them to be safe.