msbuild: Fix msbuild integration with custom installation dir and broken change detection
Variables in a PropertyGroup are set when the file is loaded, but the variables that are used there
are set later in the project settings and the variables defined in the PropertyGroup are broken. So don't
use variable declarations within PropertyGroups.
Save the file that is used to determine if the target must be executed in $(VcpkgInstalledDir) instead of
$(TLogLocation), because we don't have to rerun vcpkg.exe within all projects that use the same
$(VcpkgInstalledDir). Otherwise the target is executed for every project even is they use the same
vcpkg.json and $(VcpkgInstalledDir) and nothing changed.