[assimp] Use contrib polyclipping
It appears that the assimp port is linking polyclipping twice: once as a
dependency port via `find_package()`, and again by linking in a local copy
in `contrib/`.  This results in colliding symbols at link.

Given that assimp upstream doesn't seem to be directly compatible with
the polyclipping port version, use the one that works in `contrib/`.