[embree2] Fix static linking (#18985)
* [embree2] Fix static linking

Recently attempting to build embree2 is causing our osx workers to die. Example recent CIs:

https://dev.azure.com/vcpkg/public/_build/results?buildId=56146
https://dev.azure.com/vcpkg/public/_build/results?buildId=55886
https://dev.azure.com/vcpkg/public/_build/results?buildId=55802

Notably, we don't see this problem in PRs because it is marked "fail" in ci.baseline.txt. Initially, I was going to change it to skip, but observed that the project uses an ordinary cmake build and all the failing triplets were static ones.

If this PR passes, the hope is that CI with this change will pass too.

* Disable embree2 more directly for mac and linux because the portfile isn't prepared for it.

* Small changes

* Update the baseline version

* Use string(COMPARE EQUAL for EMBREE_STATIC_RUNTIME as well.

Co-authored-by: PhoebeHui <20694052+PhoebeHui@users.noreply.github.com>