[pistache] Delete the port.
The port "pistache" was only installing its verdored copy of rapidjson, it wasn't actually installing anything related to itself. As a result, if rapidjson was already installed by vcpkg, and the port tried to use the one provided by vcpkg, nothing at all would be installed. There is no evidence that anyone actually used this port. The upstream project does not have proper cmake bindings set up, and it will require substantial modification of their build system to repair.

As a result, this change just deletes the port. It is likely that someone could add it back later in terms of meson, which seems to be the buildsystem preferred upstream.