[openssl] fix cross compilation with mingw (#17246)
* [openssl] fix cross compilation with mingw

When cross compiling openssl on Linux, e.g.

  ./vcpkg install openssl:x64-mingw-dynamic

failed, because 'windres' was not found. The call to
openssl's Configure was missing the WINDRES environment
variable, which is needed in case it is prefixed. On
Ubuntu the binutils-mingw-w64-x86-64 package installed
it as /usr/bin/x86_64-w64-mingw32-windres.

Since mingw uses openssl-unix to build, it's worth mentioning
that the WINDRES env has no impact for non-mingw and is ignored
there.

* Handle conflicts

* Update version files

* Update ports/openssl/unix/CMakeLists.txt

* Update versions/o-/openssl.json

* Update versions/o-/openssl.json

Co-authored-by: NancyLi1013 <lirui09@beyondsoft.com>
Co-authored-by: nicole mazzuca <83086508+strega-nil-ms@users.noreply.github.com>
Co-authored-by: NancyLi1013 <46708020+NancyLi1013@users.noreply.github.com>