[OpenSSL] Allow customization of openssldir
OpenSSL searches for config file in openssldir at runtime. By default
openssldir is set to current build directory.

When an app, which uses vcpkg for building dependencies, is
distributed to another machine, searching for openssl config
in build directory on build machine doesn't make much sense.

This enables OpenSSL port users to customize this directory
with custom triplet, for example:

    set(OPENSSL_OPENSSLDIR "c:\\Program Files\\SSL")

In this case config is searched in directory which is not writable
by default.

Since vcpkg utilizes install_ssldirs target to copy config etc
to abovementioned directory, a build may fail if a directory is
read-only. To fix that, add OPENSSL_NO_INSTALL_SSLDIRS
boolean flag which makes vcpkg skip install_ssldirs target:

  set(OPENSSL_NO_INSTALL_SSLDIRS ON)

Signed-off-by: Lev Stipakov <lev@openvpn.net>