[osg] Disable conditional plugins.
Resolves nightly CI failure:
```
       "D:\buildtrees\osg\x86-windows-dbg\install.vcxproj" (default target) (1) ->
       "D:\buildtrees\osg\x86-windows-dbg\ALL_BUILD.vcxproj" (default target) (3) ->
       "D:\buildtrees\osg\x86-windows-dbg\src\osgPlugins\OpenCASCADE\osgdb_opencascade.vcxproj" (default target) (54) ->
       (ClCompile target) ->
         D:\buildtrees\osg\src\raph-3.6.5-0028e69d98.clean\src\osgPlugins\OpenCASCADE\ReaderWriterOpenCASCADE.cpp(214,49): error C2039: 'Nodes': is not a member of 'Poly_Triangulation' [D:\buildtrees\osg\x86-windows-dbg\src\osgPlugins\OpenCASCADE\osgdb_opencascade.vcxproj]
```

DICOM disabled because there's a check for dcmtk which isn't a dependency of osg[plugins]

Inventor disabled because there's a check for this library.

Opencascade disabled because it's broken.

All others disabled because they were behind some kind of if (XXX_found) in src\osgPlugins\CMakeLists.tx, except, Quicktime left alone because that's part of MacOS?