docs: CMake toolchain needs `vcpkg integrate install` first
Perhaps this is obvious to someone who has been using this tool for a
long time, but as a newcomer, I read all these sections as an either-or:
if you use MSBuild, do this, and if you use CMake, you're good to go
just using the `.toolchain` file. Well, that's not the case, CMake's
`find_path` et al won't find the *installed* packages, they also have to
be *integrated* like this.