[s2n] disable -Werror for gcc 11 build
s2n fails to build with gcc 11 (see
https://github.com/aws/s2n-tls/issues/2989) due to problematic type
declarations which generate compiler warnings, coupled with the fact
that s2n builds with -Werror by default. The underlying bug is not yet
fixed upstream.

Building with -Werror is not helpful for an end user trying to use s2n
from vcpkg, since a build failure like this is not actionable. This
change therefore uses the UNSAFE_TREAT_AS_ERRORS=OFF option to disable
-Werror.