[fontconfig] do not run fc-cache (microsoft#23920)
The fc-cache application is run on installation per default.

Font caches of the build system are useless to anyone unless
deploying fontconfig on just that build system, and even then,
the user will most likely be deploying his own fonts.conf and
correspondinly explicitly setup his users fc-cache path.

In fact, all the fc-cache run does is to generate data for
us to delete.

Instruct meson not to run fc-cache, saving us the need to
delete the cache, too.