[libunistring] Add support for building on Windows
Tweaked the portfile to no longer trigger autotools as it
caused more problems when trying to build for MSVC, so the
old automake patch is no longer needed.

The awk script used in configure should've worked, but seems
to somehow get double-interpolated somewhere along the way in
such a way as to strip out quotes needed in the script, so
put in in a dedicated script.

Added wrappers so the build scripts can use "nm" and "ar" as
if this were a UNIX target.

Fixed compiler output (warnings) getting merged with dumpbin
output and putting garbage into lib/config.h