Fix the application of VCPKG_CMAKE_CONFIGURE_OPTIONS with more than one option
VCPKG_CMAKE_CONFIGURE_OPTIONS (plural) is meant to be a list of options to be added to a cmake invocation.

But it was being applied quoted, which disallowed it from containing multiple options.
- If it was initialized with a list, cmake would get the options ";"-separated.
- If it was initialized with a blank-separated string, cmake would get the entire astring quoted as a single argument.

The fix is simple: don't (ever) quote arguments to list(APPEND) unless they are really meant to be a single element to append to the list.