Fix vcpkg.jsons that don't conform with our schema
iniparser's homepage is not a valid URI. ndevilla.free.fr/iniparser says that the homepage has moved to github, so just replaced it with that. Note that portfile.cmake already uses vcpkg_from_github.

nifticlib was referring to a github organization and used vcpkg_from_github in its homepage.

thor was missing a scheme. Using https gave an invalid certificate but http redirected to the site I put here.