[llvm] Update port to 15.0.1
The patchset is rebased onto llvmorg-15.0.1, with some significant
differences especially around install paths. The NixOS project
appears to be midway through a large change to the install path
mechanics, for the same purpose as this patchset.

When that has landed, it looks like most of these patches can go away.