[libharu] Update to 2.4.2
I updated the patches for libharu so they apply to 2.4.2 properly.
The target name changed from haru to hpdf, the haru imported
library is included for backwards compatibility. I added a usage
file that only mentions hpdf.

I added a patch for VTK so that it compiles with libharu 2.4.x,
since the type of one the HPDF_Page_SetDash arguments changed
from HPDF_UINT16 to HPDF_REAL.

POCO also has an optional libharu dependency when the 'pdf'
feature is enabled, but at the moment the PocoPDF library
doesn't seem to be built even when enabled.