Patch efsw to fix interaction with reproc on POSIX.
reproc has a limit on the number of file descriptors it attempts to
close when doing a fork/exec. This limit is hard-coded at 1024*1024
currently.
On the other hand, efsw sets RLIMIT_NOFILE to its hard maximum which on
macoS 11.6.4 is (2^63 - 1), much higher than reproc's limit.

Combined, this made every call to `reproc::process::start()` fail, which
for uplink means all the conversions using ffmpeg and gdcmconv binaries
were failing.