[FFmpeg] Fix build on ARM64
I tried to build FFmpeg for ARM64 but I got an error.
The problem happens because portfile.cmake uses gcc-9.3.0 but it has been removed from repository.
See:
https://repo.msys2.org/msys/x86_64/
So, it is just needed to update that file to the newer updated version and you will be able to build FFmpeg for ARM64 again.