[sdl2] Re-enable ibus as feature
This was disabled in #14275 because it caused the relevant libraries (ibus, glib, ...) to be depended on transitively.

Nowadays, SDL2 loads these completely at runtime, only needing the headers and such at compile time, so it's safe to re-enable this again.

Ibus is gated behind the "ibus" feature, which like the existing x11/wayland ones warn at installation that you need the necessary system packages to use it. The ibus feature is enabled by default.