[gtk] update to 4.10.1 (microsoft#30896)
gtk-4.8.2 and newer no longer contain a python script to emulate meson's
post-install logic, so we can't manually trigger it after patching meson
not to run the post-install. Patch the meson file to run the post-install
even when cross-compiling, though if it's execution is optional and all
it does is to generate some caches, i.e. pollute the build system, then
we might be better off not running the post-install logic at all.

Also, we now have a direct harfbuzz dependency by gtk/gtkmain.c and
it also needs harfbuzz with glib support enabled.