Fetch missing dependency for crashpad on linux
Note that this does not change the fact that crashpad is marked as
unsupported, but it makes the build work now with
`--allow-unsupported`.

The build seems to work fine if the default compiler is set to clang, it
does not build with gcc, which is probably why it's marked as
unsupported.