[scripts|make] Fix for Android NDK r26 (#31332)
Change how `-avoid-version` flag is passed to libtool.

Currently it is passed to configure script via LDFLAGS environment variable.
This results in it being used for configure tests which don't go through libtool.
This works with NDK r25 because passing `-avoid-version` to compiler results only in warning,
but in NDK r26 it is a hard error.
Instead pass it to Makefile after invoking configure script by overriding its LDFLAGS variable
in command line.