Use GNU Make to build autotools ports on FreeBSD (microsoft/vcpkg#32282)
For autotools ports, vcpkg runs vcpkg_configure_make to generate a GNU Makefile
project and then uses vcpkg_build_make in order to compile these Makefiles,
using the make program, which, on linux systems, is a hardlink on the gmake
program.

The BSD operating systems, however, ship with BSD make is the default make, and
GNU make, if desired, has to be installed as a third-party application.
Even if it is installed, however, GNU make will be `gmake`, while `make`
remains BSD make (cmake, among others, uses BSD Makefiles on BSD systems, so
flipping the hardlink or pushing a new make into $PATH is not a feasible
solution).

vcpkg_configure_make contains a branch for OpenBSD hosts to use gmake instead
of make (which is bmake on both OpenBSD and FreeBSD), so the most
straightforward approach to solve this issue would be to do the same on FreeBSD
systems