[bde] Remove from the curated registry.
bde has been broken in vcpkg for a long time, as when we updated to a distro where python meant python3, it failed. https://github.com/microsoft/vcpkg/pull/31961/ , merged a couple of weeks ago, fixed this problem by updating to a more recent copy of bde. Unfortunately, the new copy of bde is link-incompatible with other vcpkg ports, as they vendor ryu and pcre2 with changes.

Given that this was broken for most customers for a long time anyway, I think we should deindex for now and bring it back if/when we have an understanding with upstream as to how this is to integrate with the rest of vcpkg' ecosystem.