[capnproto] Update to v1.0.1
Updated to include capnproto f4152f83b which fixes error C7692:
"rewritten candidate function was excluded..." which occurs when
building with MSVC 2022.

The updated code brings in another bug which is only fixed in the v2
branch of capnproto (not in master which is used here). The first
patch in this commit disables the problematic code by disabling the
coroutine check in async-prelude.h

The second patch disables KJ_USE_EPOLL for Android NDK version < 23
where sigtimedwait is not defined.