fix: Make `from_gitlab` work with nested subgroups
Don't enforce a hardcoded nesting limit as these limits are enforced by
GitLab itself and may change in the future. Furthermore I believe that
users can be trusted to successfully copy'n'paste the path portion of an
URL. Therefore we just require that the assumptions of our
implementation hold, i.e. that we can convert the repo path to a CMake
list by replacing `/` with `;` and that there is a repository name.