Change z_vcpkg_fixup_rpath_in_dir to behave like other vcpkg helper functions.
This updates the list of helper port scripts in ports.cmake to ensure it has every file in scripts/cmake, and changes the caller to conditionally call the function rather than conditionally include.

This was extracted from https://github.com/microsoft/vcpkg/pull/31748/ but I'm not sure we want to do it.