Fix:
```
REGRESSION: xmlsec:x64-uwp is marked as fail but not supported for x64-uwp.
REGRESSION: yara:x64-uwp is marked as fail but not supported for x64-uwp.
REGRESSION: xmlsec:arm64-uwp is marked as fail but not supported for arm64-uwp.
REGRESSION: yara:arm64-uwp is marked as fail but not supported for arm64-uwp.
```