Allow a triplet to set `VCPKG_TARGET_` values
In `vcpkg_common_definitions.cmake` the values of `VCPKG_TARGET_` are always set. Since the file is included after the triplet file there's no way for the triplet to specify those values without them being overwritten.

Add a macro `vcpkg_set_if_not_defined` which checks if `out_var` is already `DEFINED` and if not sets the it to `default`. Use this macro in `vcpkg_common_definitions` whenever a default value is set to prevent a triplet's value from being quashed.