[libvpx] resolve bash by PATH rather then hard-coding its' location (microsoft/vcpkg#35220)
Use find_program to resolve the `bash` name.

On linux, `${MSYS_ROOT}` is undefined, so the HINT expands to `/usr/bin/bash`
which is either where `/bin/bash` points to, the location of `bash`, anyways,
or a path that is skipped while resolving the `bash` name in `${PATH}`.

As `/usr/bin` and `/bin` are both system paths, no "sane" setup will have a
`/usr/bin/bash` that is different to `/bin/bash`, so there "shouldn't" be a
behavioral change, while writing `${MSYS_ROOT}` there allows merging the line
with the WIN32 line.