[opentelemetry-cpp] Update to 1.13.0
The patch to add a dependency of opentelemetry-proto on gRPC should
not be needed after the upstream changes in:

https://github.com/open-telemetry/opentelemetry-cpp/pull/2268

Additionally, the "zpages" feature is removed as it is no longer
present upstream after having been deprecated in a previous release.

A patch is added extracted from https://github.com/open-telemetry/opentelemetry-cpp/pull/2475
and https://github.com/open-telemetry/opentelemetry-cpp/pull/2449
which fix problems that arose after `NOMINMAX` was no longer
defined within opentelemetry-cpp on Windows.

Fixes #35992.