Check for ${CC} rather then gcc
Check for `${CC}` and `${CXX}` to be installed rather then checking for `gcc`.

On Alpine, we check for `gcc` to be present, so setup `${CC}` and `${CXX}` accordingly in order to actually use the gcc that has been asserted to exist. This is nontrivial because [around line 201](https://github.com/microsoft/vcpkg/blob/13bde2ff13192e1b2fdd37bd9b475c7665ae6ae5/scripts/bootstrap.sh#L201), ${CXX}` might be set to a value that is not related to the `gcc` (distribution's default versioned) package.