Have cmake decide on a compiler if CXX is not provided
Distributors that deploy custom-named compilers also patch their cmake deployments to detect these compilers.

If multiple compilers are available on a system, have cmake decide which one to use. CMake will automatically try to use `${CXX}` if the variable is set, and otherwise decide on a compiler in a the distribution/user configurable way.