[qt5-webengine] add icu and dbus as dependencies
These are dependencies that must otherwise be provided by the
operating system. libdbus-1.so, at least on some Linux distributions,
depends on other system provided libraries like liblz4.so, libzstd.so,
and liblzma.so. It was necessary to cherry-pick two patches for the
chromium embedded in qtwebengine to make it build with ICU 74.