[qt5-webengine] Work around mixed protobuf issue
On MS Windows qt5-webengine fails to build in a context where protobuf
from vcpkg is also used because it includes the vcpkg supplied version
of protobuf headers when building the Qt part of the code in
qt5-webengine, against the internal copy of Chromium which has used
its older copy of protobuf to generate the headers. The build failure
seen is this:

    FAILED: obj/QtWebEngineCore/browser_message_filter_qt.obj
    ...
    ...
    C:\vcpkgroot\qt5-webengine\x64-windows-dbg\src\core\debug\gen\net/third_party/quiche/src/quic/core/proto/cached_network_parameters.pb.h(17): fatal error C1189: #error:  This file was generated by an older version of protoc which is

To work around this issue we ensure that the path to the chromium
provided version of protobuf is included before the path to the vcpkg
provided includes residing in:

    C:/vcpkgroot/qt5-webengine/x64-windows-dbg/include

This issue has been reported to vcpkg several times. See:

* https://github.com/microsoft/vcpkg/issues/12150
* https://github.com/microsoft/vcpkg/issues/27932
* https://github.com/microsoft/vcpkg/issues/34452