[ffmpeg] Add option to disable Truemotion1 decoder
The Truemotion1 decoder uses the truemotion1data.h file, which uses code
that was under GPL license. Although the file contains a rationale for
why the author believes this code can not be copyrighted, it can be
useful with an option to disable the decoder.

Excerpt from truemotion1data.h docstring:

    Data in this file was originally part of VpVision from On2 which is
    distributed under the GNU GPL. It is redistributed with libavcodec under
    the GNU LGPL using the common understanding that data tables necessary
    for decoding algorithms are not necessarily copyrightable.