[dbus] dbus-daemon, dbus-launch,... are first class artifacts
dbus-daemon, dbus-launch and the other executables in the dbus port
are integral components of the D-Bus system. libdbus-1 is of no
utility in itself without dbus-daemon. A case could be made that
libdbus-1 in Vcpkg could be used on Linux against the distro-provided
dbus-daemon, but this is not true on Windows and other platforms where
the dbus package is also available (and usefully so). Therefore these
executables must be treated as first-class artifacts in the vcpkg port
that can be installed along with shared libraries to a prefix by
consuming applications.

The release version of the dbus-* executables are available in the
tools folder, but that leaves a hole for dynamically linked debug
builds. These cannot install the dbus-* executables from the tools
folder as they are release binaries that are not (necessarily)
compatible with the installed libdbus-1.so/dll and other shared
library dependencies. To properly support debug builds with dynamic
linking the debug variant of these executables must also be
preserved.

Fixes #37625