[qtbase] dnslookup feature flag
Fixes QDnsLookup on !Windows.
Previously FEATURE_dnslookup was enabled but finding WrapResolv was
disabled. That builds Qt with a dummy resolver that always errors.

FEATURE_dnslookup was implicitly active before and is therefore a
default feature.
On !WINDOWS that will now require libresolv which is part of glibc.
With the dnslookup feature disabled, including QDnsLookup header will
now cause an error instead of always returning an error on actual
lookups.