[virtualgl] New port
Create a vcpkg port for [VirtualGL](https://virtualgl.org).

Creating a suggested implementation of a VirtualGL port, as there was no
response to #39936 yet.

VirtualGL will be mostly a runtime dependency, not a build-time
dependency, so in a build system such as CMake there will not be linking
of libraries but just copying of executables (such as vglrun) and
libraries (such as libvglfaker.so) to the output directory bin and lib
folder.

Closes #39936.