Added back required patches
The `fix-compiler.patch` seems to work with or without it, but it might be required on other's machines.  So I've added it back. I've also merged the `__popcnt64` fix from `fix-cmake-output.patch`.

I'm also including the highs.exe as a diagnostic tool via `vcpkg_copy_tools`.  It's quite handy, but can remove if needed.

Also added back the `fix-hconfig-path.patch`.  I had replaced it by deleting the hconfig.h file, but it was being used by a few header files.  The patch simply removes the absolute paths.