[libpng] Remove -DPNG_PREFIX=a option when apng feature enabled
This option has the effect of manipulating the ELF version and
symbol scripts applied to the shared library to define symbols like
"apng_read_end@@PNG_16_0". When using a fully shared setup and
linking other system libraries (for example, Qt6) in an application,
this causes system libraries that were linked against the system
libpng.so to fail to find libpng symbols. Because they have a
prefix of "a".

This should have been removed in the merge from two separate ports
into one port with a feature in
ff6a725392fcffa647880293e4c0a1352f0d3473