[catch2] Add feature to disable posix signal/windows seh handling
This is essential to making test crashes diagnosable. With Catch2
posix signal/windows seh handling enabled the developer will just get
the name of test and signal, which is often not enough to diagnose the
problem. With these handlers disabled the developer can instead get a
core/dmp file with all the relevant information that can be analyzed
in a debugger. See also
https://github.com/catchorg/Catch2/issues/2791.