7-zip: Only build Windows sources for Windows.
I've got no idea why these were here (it was in the initial PR without
any explanation) and can't make any sense of the upstream makefiles, but
it seems like the things in the Windows directory ought to only be built
for Windows. This fixes compilation errors with NDK r28.