Use GitHub Actions rather than Azure DevOps except macOS.
This reduces the vcpkg team's compliance burden as we won't be
maintaining a separate Azure DevOps instance, and results in a better
user experience for contributors as they need not go through so many
clicks in order to get build logs.

macOS isn't moved over yet because there's value in having all the Azure
resources on their own and that's going to be its own can of worms.

Also adds arm64-windows-static-md testing as requested by
Azure Storage internal customers.