Disable qt5-webengine on x86-windows and x64-windows
Example:

https://dev.azure.com/vcpkg/public/_build/results?buildId=109620

Azure DevOps has a hard 48 hour time limit and x86-windows and x64-windows are close to or exceeding that. qt5 is out of support from the qt foundation, and qt5-webengine is by far the most expensive port to build, at 2.1 hours.

We are concurrently investigating @autoantwort 's pipelined cache upload work to help put distance to the 48 hour limit.