[libebur128] fix build for arm64-windows
When you build libebur128 for arm64-windows, obviously that CPU doesn't support SSE2 instruction set, so a warning message is printed for telling the user that the portable C version is used instead. However, this warning made vcpkg to fail the build.
This patch simply comments that warning for the preprocessor and now the library can be built for also arm64 without troubles.