[imgui] Removed SDL2 feature bindings
Since SDL2 and SDL3 can't coexist, and vcpkg is designed to support the combination of any and all ports and features, the SDL2 and SDL3 features can't exist at the same time. Now that SDL3 is official, the decision of vcpkg team is to remove the old SDL2 bindings to support these new SDL3 ones.