Deduplicate forming the path to the exe, make args consistent:
'switch' should imply only --switch, no parameter value.
'arg' should imply that there is a parameter value expanded in place with the whole switch. Trying to splat an 'arg' is a bug.
'args' should imply that it's an array that needs to be added on the command line with splatting. Trying to not splat an 'args' is a bug.
Deduplicate forming --ci-baseline.