[qtbase] Fix path issue with Qt6CoreToolsAdditionalTargetInfo.cmake
Qt changed the path building technique in this file due to a CMake bug and never reverted it. This leads to Qt not being able to find the syncqt executable when used in Manifest mode via vcpkg, as the "tools/Qt6/" path component added by vcpkg is missing in the file.