[libjxl] Avoid manipulating CMAKE_EXE_LINKER_FLAGS for linker checks
This patches the way the linker checks are done in libjxl to avoid
manipulating CMAKE_EXE_LINKER_FLAGS, and instead use
CMAKE_REQUIRED_LINK_FLAGS as recommended in the docs. This avoids config
failures on macOS, where the -Wl,--exclude-libs flag is not supported,
when LDFLAGS contains flags that will cause the linker to fail when
-Wl,--exclude-libs=ALL is appended to it. This mostly happens with flags
such as -L/some/path.